requirements.txt
backports.functools-lru-cache==1.6.5
bitsandbytes==0.43.0
blingfire==0.1.8
blinker==1.6.2
bokeh==3.1.1
Bottleneck==1.3.5
Brotli==1.0.9
deepspeed==0.14.1
distro==1.9.0
django-cors-headers==3.11.0
django-rest-framework==0.1.0
entrypoints==0.4
fastdtw==0.3.4
FlagEmbedding==1.2.9
fschat==0.2.36
gensim==4.3.2
h5py==3.10.0
hf_transfer==0.1.6
httptools==0.6.1
hyperopt==0.2.7
immutabledict==4.1.0
insuranceqa-data==2.5.2
jieba==0.42.1
jsonlines==4.0.0
jupyter==1.0.0
jupyterlab==3.6.3
kaggle==1.6.6
lit==15.0.7
llmtuner==0.7.0
loguru==0.5.3
mkl-fft==1.3.8
mkl-random==1.2.4
modelscope==1.13.3
mpi4py==3.1.6
munkres==1.1.4
netron==7.4.6
nltk==3.2.5
openai==0.28.0
openpyxl==3.1.2
pdfplumber==0.11.0
pipdeptree==2.20.0
PyJWT==2.4.0
PyMySQL==1.0.2
pyOpenSSL==23.2.0
PyQt5==5.15.10
PySocks==1.7.1
pytest==8.1.1
python-dotenv==1.0.1
reformer-pytorch==1.4.4
responses==0.13.3
retrying==1.3.4
rouge-chinese==1.0.3
SciencePlots==2.1.0
seaborn==0.12.2
sip==6.7.12
statsmodels==0.14.0
tables==3.8.0
tensorboardX==2.6.2.2
torch_geometric==2.4.0
torch-tb-profiler==0.4.3
torchaudio==2.0.2+cu118
torchdiffeq==0.2.3
torchvision==0.15.2+cu118
transformers-stream-generator==0.0.5
uvloop==0.19.0
vllm==0.4.1
watchfiles==0.21.0
wavedrom==2.0.3.post3
wheel==0.41.2
